affichage si aucun message

+++ view/forum/listMessages.php
@@ -33,32 +33,37 @@ if(App\Session::getVisiteur()){
 
 <div class="listeMessages">
     <?php
-    foreach($messages as $message){
-        ?>            
-            <div class="message2">
-                <a class="pseudo" href="index.php?ctrl=forum&action=profilVisiteur&id=<?=$message->getVisiteur()->getId() ?>"><?=$message->getVisiteur()->getPseudonyme()?></a>
-                <p class="role"><?=$message->getVisiteur()->getRole()?></p>
-                <p class="date"><?=$message->getDateCreation()?></p>
-                <p class="texte"><?=$message->getMessage()?></p>
-                <div class="interactions">
-                    <?php if(App\Session::getVisiteur()){ ?>
-                    <div class="vote">
-                        <a href="index.php?ctrl=forum&action=upvoteMessage&id=<?= $message->getId()?>"><i class="fa-solid fa-up-long"></i></a>
-                        <p><?=$message->getNbVote()?></p>
-                        <a href="index.php?ctrl=forum&action=downvoteMessage&id=<?= $message->getId()?>"><i class="fa-solid fa-down-long"></i></a>
+    if(isset($messages)){
+        foreach($messages as $message){
+            ?>            
+                <div class="message2">
+                    <a class="pseudo" href="index.php?ctrl=forum&action=profilVisiteur&id=<?=$message->getVisiteur()->getId() ?>"><?=$message->getVisiteur()->getPseudonyme()?></a>
+                    <p class="role"><?=$message->getVisiteur()->getRole()?></p>
+                    <p class="date"><?=$message->getDateCreation()?></p>
+                    <p class="texte"><?=$message->getMessage()?></p>
+                    <div class="interactions">
+                        <?php if(App\Session::getVisiteur()){ ?>
+                        <div class="vote">
+                            <a href="index.php?ctrl=forum&action=upvoteMessage&id=<?= $message->getId()?>"><i class="fa-solid fa-up-long"></i></a>
+                            <p><?=$message->getNbVote()?></p>
+                            <a href="index.php?ctrl=forum&action=downvoteMessage&id=<?= $message->getId()?>"><i class="fa-solid fa-down-long"></i></a>
+                        </div>
+                        <?php
+                            if(App\Session::isAdmin() || App\Session::getVisiteur()->getId() == $message->getVisiteur()->getId()){ ?>
+                                <div class="supprimer">
+                                    <a class="btnSupprimer" href="index.php?ctrl=forum&action=supprimerMessage&id=<?= $message->getId()?>">Supprimer</a>
+                                </div>
+                            <?php 
+                            }
+                        } ?>
                     </div>
-                    <?php
-                        if(App\Session::isAdmin() || App\Session::getVisiteur()->getId() == $message->getVisiteur()->getId()){ ?>
-                            <div class="supprimer">
-                                <a class="btnSupprimer" href="index.php?ctrl=forum&action=supprimerMessage&id=<?= $message->getId()?>">Supprimer</a>
-                            </div>
-                        <?php 
-                        }
-                    } ?>
-                </div>
-            </div>            
-        <?php
-    }
+                </div>            
+            <?php
+        }
+    }else{ ?>
+        <p>Aucun message</p>
+    <?php
+    } 
 ?>
